+++ core/services/chainlink/application.go
@@ -20,7 +20,6 @@ import (
 	"github.com/smartcontractkit/chainlink/core/services/pipeline"
 	"github.com/smartcontractkit/chainlink/core/services/postgres"
 	"github.com/smartcontractkit/chainlink/core/services/synchronization"
-	"github.com/smartcontractkit/chainlink/core/services/telemetry"
 	strpkg "github.com/smartcontractkit/chainlink/core/store"
 	"github.com/smartcontractkit/chainlink/core/store/models"
 	"github.com/smartcontractkit/chainlink/core/store/orm"
@@ -85,7 +84,6 @@ type ChainlinkApplication struct {
 	shutdownOnce             sync.Once
 	shutdownSignal           gracefulpanic.Signal
 	balanceMonitor           services.BalanceMonitor
-	monitoringEndpoint       telemetry.MonitoringEndpoint
 	explorerClient           synchronization.ExplorerClient
 }
 
@@ -100,12 +98,10 @@ func NewApplication(config *orm.Config, ethClient eth.Client, advisoryLocker pos
 
 	explorerClient := synchronization.ExplorerClient(&synchronization.NoopExplorerClient{})
 	statsPusher := synchronization.StatsPusher(&synchronization.NoopStatsPusher{})
-	telemetryAgent := telemetry.MonitoringEndpoint(&telemetry.NoopAgent{})
 
 	if config.ExplorerURL() != nil {
 		explorerClient = synchronization.NewExplorerClient(config.ExplorerURL(), config.ExplorerAccessKey(), config.ExplorerSecret())
-		statsPusher = synchronization.NewStatsPusher(store.ORM, explorerClient)
-		telemetryAgent = telemetry.NewAgent(explorerClient)
+		statsPusher = synchronization.NewStatsPusher(store.DB, explorerClient)
 	}
 
 	runExecutor := services.NewRunExecutor(store, statsPusher)
@@ -113,6 +109,7 @@ func NewApplication(config *orm.Config, ethClient eth.Client, advisoryLocker pos
 	runManager := services.NewRunManager(runQueue, config, store.ORM, statsPusher, store.Clock)
 	jobSubscriber := services.NewJobSubscriber(store, runManager)
 	gasUpdater := services.NewGasUpdater(store)
+	promReporter := services.NewPromReporter(store.DB.DB())
 	logBroadcaster := eth.NewLogBroadcaster(ethClient, store.ORM, store.Config.BlockBackfillDepth())
 	eventBroadcaster := postgres.NewEventBroadcaster(config.DatabaseURL(), config.DatabaseListenerMinReconnectInterval(), config.DatabaseListenerMaxReconnectDuration())
 	fluxMonitor := fluxmonitor.New(store, runManager, logBroadcaster)
@@ -159,7 +156,6 @@ func NewApplication(config *orm.Config, ethClient eth.Client, advisoryLocker pos
 		pendingConnectionResumer: pendingConnectionResumer,
 		shutdownSignal:           shutdownSignal,
 		balanceMonitor:           balanceMonitor,
-		monitoringEndpoint:       telemetryAgent,
 		explorerClient:           explorerClient,
 	}
 
@@ -171,6 +167,7 @@ func NewApplication(config *orm.Config, ethClient eth.Client, advisoryLocker pos
 		jobSubscriber,
 		pendingConnectionResumer,
 		balanceMonitor,
+		promReporter,
 	)
 
 	for _, onConnectCallback := range onConnectCallbacks {